+++ src/chart/rendering/drawlabel.cpp
@@ -4,98 +4,73 @@ namespace Vizzu::Draw
 {
 
 void DrawLabel::draw(Gfx::ICanvas &canvas,
-    const Geom::TransformedRect &rect,
+    const Geom::TransformedRect &fullRect,
     const std::string &text,
     const Styles::Label &style,
     Util::EventDispatcher::Event &onDraw,
     std::unique_ptr<Util::EventTarget> eventTarget,
     Options options) const
 {
-	auto relRect = Geom::Rect{Geom::Point(), rect.size};
+	auto relativeRect = Geom::Rect{{}, fullRect.size};
 
 	if (!style.backgroundColor->isTransparent()) {
 		canvas.save();
-		canvas.setBrushColor(*style.backgroundColor);
-		canvas.setLineColor(*style.backgroundColor);
-		canvas.transform(rect.transform);
-		canvas.rectangle(relRect);
+		auto bgColor = *style.backgroundColor * options.bgAlpha;
+		canvas.setBrushColor(bgColor);
+		canvas.setLineColor(bgColor);
+		canvas.transform(fullRect.transform);
+		canvas.rectangle(relativeRect);
 		canvas.restore();
 	}
 
+	auto font = Gfx::Font{style};
+	auto paddedRect = style.contentRect(relativeRect, font.size);
+	auto [alignRect, alignConstant] = alignText(paddedRect,
+	    style,
+	    Gfx::ICanvas::textBoundary(font, text));
+
 	canvas.save();
 
-	auto contentRect =
-	    style.contentRect(relRect, style.calculatedSize());
-
-	auto font = Gfx::Font{style};
 	canvas.setFont(font);
-	if (options.setColor)
-		canvas.setTextColor(*style.color * options.alpha);
+	if (options.alpha)
+		canvas.setTextColor(*style.color * *options.alpha);
 
-	auto textSize = Gfx::ICanvas::textBoundary(font, text);
-	auto alignSize = textSize;
-	alignSize.x = std::min(alignSize.x, contentRect.size.x);
-	auto textRect = alignText(contentRect, style, alignSize);
-	textRect.size = textSize;
+	if (onDraw.invoke(Events::OnTextDrawEvent{*eventTarget,
+	        fullRect,
+	        paddedRect,
+	        alignConstant,
+	        text})) {
 
-	auto transform =
-	    rect.transform * Geom::AffineTransform(textRect.bottomLeft());
+		auto textTransform =
+		    fullRect.transform
+		    * Geom::AffineTransform(alignRect.bottomLeft());
 
-	if (options.flip)
-		transform = transform
-		          * Geom::AffineTransform(textRect.size, 1.0, -M_PI);
+		if (options.flip)
+			textTransform *=
+			    Geom::AffineTransform(alignRect.size, 1.0, -M_PI);
 
-	Geom::TransformedRect trRect;
-	trRect.transform = transform;
-	trRect.size = textRect.size;
+		canvas.transform(textTransform);
 
-	if (onDraw.invoke(
-	        Events::OnTextDrawEvent{*eventTarget, trRect, text})) {
-		canvas.transform(transform);
+		canvas.text({{}, alignRect.size}, text);
 
-		canvas.text(Geom::Rect(Geom::Point(), textRect.size), text);
-
-		renderedChart.emplace(trRect, std::move(eventTarget));
+		renderedChart.emplace(fullRect, std::move(eventTarget));
 	}
 
 	canvas.restore();
 }
 
-double DrawLabel::getHeight(const Styles::Label &style)
-{
-	const Gfx::Font font(style);
-	auto textHeight = Gfx::ICanvas::textBoundary(font, "").y;
-	return style.paddingTop->get(textHeight, font.size)
-	     + style.paddingBottom->get(textHeight, font.size)
-	     + textHeight;
-}
-
-Geom::Rect DrawLabel::alignText(const Geom::Rect &contentRect,
+std::pair<Geom::Rect, double> DrawLabel::alignText(
+    const Geom::Rect &paddedRect,
     const Styles::Label &style,
     const Geom::Size &textSize)
 {
-	Geom::Rect res;
-	res.size = textSize;
-	res.pos = contentRect.pos;
+	Geom::Rect res{paddedRect.pos, textSize};
 
-	res.pos.x = style.textAlign->combine<double>(
-	    [&contentRect, &textSize](int,
-	        const Styles::Text::TextAlign &align) -> double
-	    {
-		    switch (align) {
-		    case Styles::Text::TextAlign::left:
-			    return contentRect.left();
+	auto align = style.textAlign->calculate<double>();
+	if (auto space = paddedRect.size.x - textSize.x; space > 0)
+		res.pos.x += space / 2.0 * (1 + align);
 
-		    case Styles::Text::TextAlign::right:
-			    return contentRect.right() - textSize.x;
-
-		    default:
-		    case Styles::Text::TextAlign::center:
-			    return contentRect.center().x - textSize.x / 2.0;
-		    }
-	    });
-
-	return res;
+	return {res, align};
 }
 
 }