+++ src/chart/rendering/orientedlabel.h
@@ -1,6 +1,8 @@
 #ifndef DRAWORIENTEDLABEL_H
 #define DRAWORIENTEDLABEL_H
 
+#include <optional>
+
 #include "base/geom/transformedrect.h"
 #include "chart/rendering/drawlabel.h"
 
@@ -9,23 +11,15 @@
 namespace Vizzu::Draw
 {
 
-class OrientedLabel
+struct OrientedLabel : DrawLabel
 {
-public:
-	const std::string &text;
-	Geom::TransformedRect rect{};
-	Geom::Rect contentRect{};
-
-	[[nodiscard]] static OrientedLabel create(Gfx::ICanvas &canvas,
+	void draw(Gfx::ICanvas &canvas,
 	    const std::string &text,
 	    const Geom::Line &labelPos,
 	    const Styles::OrientedLabel &labelStyle,
-	    double centered);
-
-	void draw(Gfx::ICanvas &canvas,
-	    RenderedChart &renderedChart,
-	    const Gfx::Color &textColor,
-	    const Gfx::Color &bgColor,
+	    double centered,
+	    std::optional<double> textAlpha,
+	    double bgAlpha,
 	    Util::EventDispatcher::Event &event,
 	    std::unique_ptr<Util::EventTarget> eventTarget) const;
 };